Convert some static sets to frozensets (#671)
* Convert some static sets to frozensets

It does not affect performance much, but as far as
I can tell this should reduce threads contention and
maybe increase performance a bit on Python 3.15t+ as
on frozenset it might be able to not held some extra locks
for frozensets.

On my local test this was less than 0.5% faster on
free-threaded as everything was dwarfed by the time
lru_cache locks; but those pesky lines in the profiler
were bothering me.

* :art: reformat constant.py

---------

Co-authored-by: Ahmed TAHRI <tahri.ahmed@proton.me>